- WigiiCoreExecutor: on group portal url (link group to web address), supports file://path/to/folder/on/disk protocol to automatically synchronize Wigii elements with files found on server file system - added new configuration parameter isUniqueInGroup which prevents duplicates in same group only, ignoring check on sub-groups. This is useful for example to emulate a file system in which file names are unique in same folder only. - added function resolveFilePath which resolves a Wigii file path on the server and returns a physical path from which the file can be read. Replaced usage of FILES_PATH constant with a call to this function. - supports file://path/to/folder/on/disk protocol for Wigii files path. In that case the file is searched into a location on server under the CLIENT_DATA_PATH folder. - WigiiBPL: added method groupSyncElementsWithFileSystem which synchronizes the elements contained in a group with the files present in a physical folder on the file system

+++ data/wigiiSystem/core/_webImplExecutor/libs/form/FormExecutor.php
@@ -766,10 +766,12 @@ abstract class FormExecutor extends Model implements RecordStructureFactory, TRM
 			// Detects Box files
 			$isBoxFile = $fileName && strstr($fileName, "box://");
 			$isOldBoxFile = $oldFileName && strstr($oldFileName, "box://");
+			// Detects client files
+			$isOldClientFile = $oldFileName && strstr($oldFileName, "file://");
 			
 			//if the deleteFile is checked then delete the oldFile + thumbnail
 			//this work only if oldRecord is defined
-			if($fileName == null && $oldRecord!=null && !$isOldBoxFile){
+			if($fileName == null && $oldRecord!=null && !$isOldBoxFile && !$isOldClientFile){
 				if (isImage($oldRecord->getFieldValue($fieldName, "mime"))) @unlink(FILES_PATH."tn_".$oldFileName);
 				if($oldFileName && $fieldXml["keepHistory"]>0){
 					$this->pushFileToHistory($p, $exec, (string)$fieldXml["keepHistory"], $oldRecord, $fieldName);
@@ -785,7 +787,6 @@ abstract class FormExecutor extends Model implements RecordStructureFactory, TRM
 				$filePath = TEMPORARYUPLOADEDFILE_path.$fileName;
 		
 				if(!$storeFileInWigiiBag){
-					//						fput("move  $filePath in ".FILES_PATH.$fileName);
 					//move the tempFile into client folder
 					if(!rename($filePath, FILES_PATH.$fileName)) throw new ServiceException("Error on storing the temporaryUploadedFile ".$filePath." in the Client folder:".FILES_PATH.$fileName, ServiceException::FORBIDDEN);
 		
@@ -804,12 +805,11 @@ abstract class FormExecutor extends Model implements RecordStructureFactory, TRM
 					}
 				}
 		
-				//					fput("delete old file $fieldName:".$oldFileName);
 				//delete old file if setted
 				if($oldFileName != null && !$isOldBoxFile && isImage($oldRecord->getFieldValue($fieldName, "mime"))){
 					@unlink(FILES_PATH."tn_".$oldFileName);
 				}
-				if($oldFileName != null && !$isOldBoxFile){
+				if($oldFileName != null && !$isOldBoxFile && !$isOldClientFile){
 					if($oldFileName && $fieldXml["keepHistory"]>0){
 						$this->pushFileToHistory($p, $exec, (string)$fieldXml["keepHistory"], $oldRecord, $fieldName);
 					} else {
